ref: migrate -> `LootGranter`
