+++ src/Imlight.CoreLib/Game/Results/Handlers/ResDropTableHandler.cs
@@ -17,23 +17,17 @@
  */
 
 using System;
-using System.Collections.Generic;
 using Akka.Actor;
-using Imcodec.MessageLayer.Generated;
-using Imcodec.ObjectProperty;
 using Imcodec.ObjectProperty.TypeCache;
 using Imlight.Common;
 using Imlight.CoreLib.Game.DropTables;
 using Imlight.CoreLib.Shared.Packets;
-using Imlight.CoreLib.WizardData.Models.Player;
-using Imlight.CoreLib.WizardData.Models.World;
 
 namespace Imlight.CoreLib.Game.Results.Handlers;
 
 internal sealed class ResDropTableHandler : BaseResultHandler<ResDropTable> {
 
     private const float QUERY_WIZARD_TIMEOUT_SECONDS = 5.0f;
-    private const uint LOOT_LIST_SERIALIZATION_FLAGS = 4;
 
     public override bool Execute(IResultContext context) {
         // Context does not ship with a wizard reference, so we need to query for it.
@@ -64,127 +58,9 @@ internal sealed class ResDropTableHandler : BaseResultHandler<ResDropTable> {
                                                wizard);
 
         // Process the results.
-        UpdateWizardGold(context.GetPlayerRef(), wizard, rollResults.GoldAmount);
-        UpdateWizardXP(context.GetPlayerRef(), rollResults.ExperienceAmount);
-        UpdateWizardTP(context.GetPlayerRef(), wizard, rollResults.TrainingPoints);
-        UpdateCharacterItems(wizard, rollResults.Items);
-        SendLootInfoToClient(context.GetPlayerRef(), rollResults, wizard);
-
-        if (rollResults.GrantsPotionSlot) {
-            UpdateWizardPotionMax(context.GetPlayerRef(), wizard);
-        }
+        LootGranter.GrantAndDisplay(context.GetPlayerRef(), wizard, rollResults);
 
         return true;
-    }
-
-    private static void UpdateWizardGold(IActorRef playerActor, Wizard wizard, int goldDelta) {
-        if (goldDelta == 0) {
-            return;
-        }
-
-        // Add gold to the wizard. This will save their data, but not inform their game client.
-        wizard.AddGold(goldDelta);
-
-        // Now, inform the game client their gold has been updated.
-        // This only changes the character page. It does not show a popup or anything.
-        // The popup comes from the network LootInfoList.
-        var networkMessage = new WIZARD_12_PROTOCOL.MSG_UPDATEGOLD() {
-            Gold = wizard.GameStats.m_currentGold,
-            MaxGold = wizard.GameStats.m_baseGoldPouch
-        };
-        playerActor.Tell(networkMessage);
-    }
-
-    private static void UpdateWizardXP(IActorRef playerActor, int xpDelta) {
-        if (xpDelta == 0) {
-            return;
-        }
-
-        // XP is simple and we only need to inform the SessionActor.
-        var internalMsg = new CHARACTER_103_PROTOCOL.MSG_GAINXP {
-            XP = xpDelta
-        };
-        playerActor.Tell(internalMsg);
-
-        // That's it. The SessionActor will inform the game client, and level them up if needed.
-    }
-
-    private static void UpdateWizardTP(IActorRef playerActor, Wizard wizard, int tpDelta) {
-        if (tpDelta == 0) {
-            return;
-        }
-
-        // Set the new amount of training points for the wizard.
-        var oldTP = wizard.MagicSchoolBehavior.TrainingPoints;
-        var newTP = Math.Max(0, oldTP + tpDelta);
-        if (newTP == oldTP) {
-            return;
-        }
-
-        wizard.UpdateTrainingPoints(newTP);
-
-        // Inform the game client of the new TP amount.
-        var msg = new WIZARD_12_PROTOCOL.MSG_UPDATETRAINING() {
-            TrainingPoints = (ushort) newTP
-        };
-        playerActor.Tell(msg);
-    }
-
-    private static void UpdateCharacterItems(Wizard wizard, List<DropItemResult> items) {
-        if (items.Count == 0) {
-            return;
-        }
-
-        // Add each item to the wizard's inventory.
-        foreach (var item in items) {
-            if (!ulong.TryParse(item.ItemId, out var itemGuid)) {
-                continue;
-            }
-
-            if (!wizard.AddItemToInventory(itemGuid, out var _)) {
-                Logger.Error("Handler failed to add item {0} to wizard {1}'s inventory.",
-                    Logger.Args(item.ItemId, wizard.CharId));
-
-                continue;
-            }
-        }
-    }
-
-    private static void UpdateWizardPotionMax(IActorRef playerActor, Wizard wizard) {
-        var currentWizardMaxPots = wizard.GameStats.m_potionMax;
-        var newWizardMaxPots = currentWizardMaxPots + 1;
-
-        wizard.UpdatePotions(newWizardMaxPots, newWizardMaxPots);
-
-        // Inform the player's game client that their potion charge has been updated.
-        var potionChargeUpdateMsg = new WIZARD_12_PROTOCOL.MSG_UPDATEPOTIONS {
-            PotionMax = newWizardMaxPots,
-            PotionCharge = newWizardMaxPots
-        };
-        playerActor.Tell(potionChargeUpdateMsg);
-    }
-
-    private static void SendLootInfoToClient(IActorRef playerActor, DropTableResult results, Wizard wizard) {
-        // Inform the game client of the loot results.
-        if (!DropTableConverter.HasRewards(results)) {
-            return;
-        }
-
-        // Convert the loot results into something we can send over the network.
-        var networkLootList = DropTableConverter.ToLootInfoList(results);
-        var serializer = new ObjectSerializer(Versionable: false);
-        if (!serializer.Serialize(networkLootList, LOOT_LIST_SERIALIZATION_FLAGS, out var serializedLootList)) {
-            Logger.Error("Handler failed to serialize loot list for network transmission.");
-
-            return;
-        }
-
-        var lootMsg = new WIZARD_12_PROTOCOL.MSG_LOOT() {
-            GlobalID = wizard.CharId,
-            LootList = serializedLootList
-        };
-
-        playerActor.Tell(lootMsg);
     }
 
 }
